chore: restructure src/server into domain folders for auth0-server-js migration
