+++ src/client/mfa/mfa-client.flow.test.ts
@@ -11,6 +11,7 @@ import {
   vi
 } from "vitest";
 
+import { encryptMfaToken } from "../../server/mfa/mfa-utils.js";
 import {
   challengeScenariosClient as challengeScenarios,
   enrollScenarios,
@@ -19,7 +20,6 @@ import {
 } from "../../test-fixtures/mfa-scenarios-shared.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import type { MfaClient } from "../../types/index.js";
-import { encryptMfaToken } from "../../utils/mfa-utils.js";
 
 // Test constants
 const DEFAULT = {
