+++ src/server/passkey/server-passkey-client.test.ts
@@ -18,7 +18,7 @@ import {
 } from "../../test-fixtures/defaults.js";
 import { AuthClientProvider } from "../auth-client-provider.js";
 import { AuthClient } from "../auth-client.js";
-import { encrypt } from "../cookies.js";
+import { encrypt } from "../cookies/index.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
 import { ServerPasskeyClient } from "./server-passkey-client.js";
