+++ src/server/session/stateless-session-store.ts
@@ -5,7 +5,7 @@ import {
   CookieOptions,
   SessionData
 } from "../../types/index.js";
-import * as cookies from "../cookies.js";
+import * as cookies from "../cookies/index.js";
 import {
   AbstractSessionStore,
   BeforeSessionRolledHook,
