+++ src/server/tests/auth-client.proxy.test.ts
@@ -17,7 +17,7 @@ import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import { SessionData } from "../../types/index.js";
 import { AuthClient } from "../auth-client.js";
-import { decrypt, encrypt } from "../cookies.js";
+import { decrypt, encrypt } from "../cookies/index.js";
 import { generateDpopKeyPair } from "../dpop/retry.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
