+++ src/server/tests/logout-strategy.flow.test.ts
@@ -15,7 +15,7 @@ import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import type { SessionData } from "../../types/index.js";
 import { AuthClient } from "../auth-client.js";
-import { encrypt } from "../cookies.js";
+import { encrypt } from "../cookies/index.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
