+++ src/server/tests/session-expiry.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, it } from "vitest";
 
 import { AccessTokenErrorCode } from "../../errors/oauth-errors.js";
-import { getSessionChangesAfterGetAccessToken } from "../../utils/session-changes-helpers.js";
+import { getSessionChangesAfterGetAccessToken } from "../session/session-changes-helpers.js";
 import {
   buildSessionFromCallback,
   isSessionCeilingInPast,
   isSessionCeilingReached
-} from "../../utils/session-helpers.js";
+} from "../session/session-helpers.js";
 
 // ---------------------------------------------------------------------------
 // isSessionCeilingReached
