+++ src/test-fixtures/mocks.ts
@@ -1,7 +1,7 @@
 import * as oauth from "oauth4webapi";
 import { vi } from "vitest";
 
-import { RequestCookies } from "../server/cookies.js";
+import { RequestCookies } from "../server/cookies/index.js";
 
 /**
  * Creates a mock factory for next/headers.js
