+++ src/testing/generate-session-cookie.test.ts
@@ -1,7 +1,7 @@
 import * as jose from "jose";
 import { describe, expect, it } from "vitest";
 
-import { decrypt } from "../server/cookies.js";
+import { decrypt } from "../server/cookies/index.js";
 import { generateSecret } from "../test-fixtures/utils.js";
 import { SessionData } from "../types/index.js";
 import {
