+++ src/testing/generate-session-cookie.ts
@@ -1,4 +1,4 @@
-import { encrypt } from "../server/cookies.js";
+import { encrypt } from "../server/cookies/index.js";
 import { SessionData } from "../types/index.js";
 
 export type GenerateSessionCookieConfig = {
